pyproject.toml
[tool.poetry]
name = "mock-server"
version = "1.0.0"
description = "mock-server for front-end project"
authors = ["punkyoon <punkkid001@gmail.com>"]
license = "Unlicense"

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.2"
djangorestframework = "^3.9"
django-safedelete = "^0.5.1"
django-model-utils = "^3.1"
djangorestframework-serializer-extensions = "^1.0"
django-countries = "^5.3"
django-rest-auth = "^0.9.5"
djangorestframework-jwt = "^1.11"
django-cors-headers = "^3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
